% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivePeople-data.R
\docType{data}
\name{ActivePeople}
\alias{ActivePeople}
\title{Economically active people in Spain from 2002 to 2019.}
\format{A data frame with columns:
\describe{
 \item{Time}{A trimester (T1, T2, T3 and T4) between 2002 and 2019.}
 \item{TotalWomen}{Number (thousands) of economically active women.}
 \item{TotalMen}{Number (thousands) of economically active men.}
}}
\source{
Spanish Statistical Office (INE), \url{http://www.ine.es}
}
\usage{
data(ActivePeople)
}
\description{
Number (thousands) of economically active women and men in Spain between 2002 and 2019
}
\examples{
\dontrun{
 ActivePeople
}
}
\keyword{datasets}
