% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mortgages-data.R
\docType{data}
\name{Mortgages}
\alias{Mortgages}
\title{Mortgages constituted on urban properties in Spain from 2003 to 2018.}
\format{A data frame with columns:
\describe{
  \item{Year}{Year from 2003 to 2018}
  \item{Number.of.bank.mortgages}{Numbers of bank mortages.}
  \item{Amount.of.bank.mortgages}{Amount (thousands of euros) of bank mortages.}
  \item{Number.of.savings.bank.mortgages}{Numbers of savings bank mortages.}
  \item{Amount.of.savings.bank.mortgages}{Amount (thousands of euros) of savings bank mortages.}
  \item{Number.of.other.entities.mortgages}{Numbers of other entities mortages.}
  \item{Amount.of.other.entities.mortgages}{Amount (thousands of euros) of other entities mortages.}
}}
\source{
Spanish Statistical Office (INE), \url{http://www.ine.es}
}
\usage{
data(Mortgages)
}
\description{
Number of mortgages constituted on urban properties and mortgages amounts (thousands of euros) from 2003 to 2018.
}
\examples{
\dontrun{
 Mortgages
}
}
\keyword{datasets}
